Add agent memory to one-shot agent
